# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Car Reservation API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/car": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Ambil semua data mobil dengan filter opsional",
                "operationId": "c5e4dc731be2f3d08b7f82918e701076",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Filter berdasarkan user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filter berdasarkan nama mobil",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "brand_name",
                        "in": "query",
                        "description": "Filter berdasarkan merek mobil",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price_min",
                        "in": "query",
                        "description": "Harga minimum sewa per hari",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "price_max",
                        "in": "query",
                        "description": "Harga maksimum sewa per hari",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "stock_min",
                        "in": "query",
                        "description": "Stock minimal",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "stock_max",
                        "in": "query",
                        "description": "Stock maksimal",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Semua data mobil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cars": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CarResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Car"
                ],
                "summary": "Tambah mobil baru",
                "operationId": "1269980bf78594760ded521eaaf997dc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "category_id",
                                    "name",
                                    "brand_name",
                                    "price_per_day",
                                    "stock"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Avanza"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "avanza.jpg"
                                    },
                                    "brand_name": {
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "price_per_day": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 350000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mobil berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Terjadi kesalahan saat menambahkan mobil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/car/category/{categoryId}": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Ambil mobil berdasarkan kategori",
                "operationId": "0f1fb758c798f59bafb4756354365deb",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID kategori",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar mobil berdasarkan kategori",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CarResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ada mobil ditemukan"
                    }
                }
            }
        },
        "/api/car/{id}": {
            "get": {
                "tags": [
                    "Car"
                ],
                "summary": "Ambil detail mobil",
                "operationId": "e733042a1d10af8eca4b462c093828de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID mobil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data mobil berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mobil tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Car"
                ],
                "summary": "Update data mobil",
                "operationId": "8e2ad7fd7c5c5d9d99ad90929d46f960",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID mobil yang akan diupdate",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Avanza Updated"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "avanza_new.jpg"
                                    },
                                    "brand_name": {
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "price_per_day": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 370000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data mobil berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CarResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mobil tidak ditemukan"
                    },
                    "500": {
                        "description": "Terjadi kesalahan saat memperbarui data mobil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Car"
                ],
                "summary": "Hapus mobil",
                "operationId": "e28d13360bbfab4a4e50d215c4342662",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID mobil",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mobil berhasil dihapus"
                    },
                    "404": {
                        "description": "Mobil tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CarResource": {
                "title": "Car Resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Avanza"
                    },
                    "brand_name": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 350000
                    },
                    "type": {
                        "type": "string",
                        "example": "SUV"
                    },
                    "transmission": {
                        "type": "string",
                        "example": "automatic"
                    },
                    "fuel": {
                        "type": "string",
                        "example": "petrol"
                    },
                    "seat": {
                        "type": "integer",
                        "example": 5
                    },
                    "image": {
                        "type": "string",
                        "example": "car.jpg"
                    },
                    "available": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication using Laravel Sanctum",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "security": [
                {
                    "sanctum": []
                }
            ]
        }
    },
...
}